Fixed a bug in fu's handling of HEAD requests.
The fu.js code wasn't checking to see if a request method was HEAD.
That resulted in it sending a body to a head request.  This in turn
resulted in breaking the spec which the underlying Node library didn't
like.  In turn, this resulted in an uncaught exception which crashed the
server. Now if the request type is HEAD, it does the right thing.